pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-gateway-bench"
version = "0.2.0"
description = "CLI benchmarking tool for LLM API gateways — measure latency, TTFT, and throughput across providers"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["llm", "benchmark", "openai", "anthropic", "latency", "ttft", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "anyio>=4.0.0",
    "anthropic>=0.30.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "mypy>=1.8.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "types-PyYAML>=6.0.12",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocs-git-revision-date-localized-plugin>=1.3.0",
    "pymdown-extensions>=10.8.0",
]

[project.scripts]
lgb = "llm_gateway_bench.cli:app"

[project.urls]
Homepage = "https://github.com/mnbplus/llm-gateway-bench"
Docs = "https://mnbplus.github.io/llm-gateway-bench/"
Changelog = "https://github.com/mnbplus/llm-gateway-bench/releases"
Repository = "https://github.com/mnbplus/llm-gateway-bench"
Issues = "https://github.com/mnbplus/llm-gateway-bench/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_gateway_bench"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
python_version = "3.9"
pretty = true
show_error_codes = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true
mypy_path = "src"

[tool.pytest.ini_options]
addopts = "-ra --cov=llm_gateway_bench --cov-report=term-missing --cov-report=xml"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["llm_gateway_bench"]
omit = [
    "tests/*",
    "*/__main__.py",
    "*/site-packages/*",
    "*/.venv/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == \"__main__\":",
]
